+++ CHANGELOG.md
@@ -1,5 +1,12 @@
-# v2019.8.15
+# CHANGELOG
+
+## v2019.8.15
 
 - primary functionality implemented
 - generate secure random passwords and passphrases
 - click to copy
+
+## v2022.09.06
+
+- migrate from Create React App to Vite
+- bump react-spring to latest version